chore: rename other classes
